requirements.txt
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Cython==0.29.28
docopt==0.6.2
ghp-import==2.0.2
idna==3.3
importlib-metadata==4.11.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
mergedeep==1.3.4
mkdocs==1.2.3
mkdocs-material==8.2.1
mkdocs-material-extensions==1.0.3
mypy-extensions==0.4.3
numpy==1.22.2
packaging==21.3
Panda3D==1.10.11.post1
panda3d-gltf==0.13
panda3d-simplepbr==0.9
pathspec==0.9.0
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
Pygments==2.11.2
pymdown-extensions==9.2
pyobjc-core==8.2
pyobjc-framework-Cocoa==8.2
pyparsing==3.0.7
pyperclip==1.8.2
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.27.1
screeninfo==0.8
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
ursina==4.1.1
watchdog==2.1.6
yarg==0.1.9
zipp==3.7.0
